refactor: rename normalised-task-status-chart to task-status-summary-charts
