+++ src/app/units/states/analytics/unit-analytics-route.component.html
@@ -23,7 +23,7 @@
         }
       </mat-card-content>
     </mat-card>
-    <f-normalised-task-status-chart [unit]="unit" />
+    <f-task-status-summary-charts [unit]="unit" />
     <f-analytics-tutor-times [downloadCsvFn]="downloadCsv" [unit]="unit"></f-analytics-tutor-times>
   </div>
 </div>
